Add files via upload
Library Members can avail of the following facilities: 
1. Providing books/journals for reading/reference. 
2. Issue of books to students so that they can take them home and return them on or before the 
due date. When a new book is brought to the library, its record is created, which can be modified 
or deleted if the book is no longer available in the library. If a student requests for a book, it is 
first searched for its availability through a unique key field “Bno”, i.e., Book code. If it is 
available (i.e., not issued to any other member) then it is issued and details are entered in the 
Issue module.